+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.5] - 18-10-2025
 
+### Fix
+
+- fix clearing of quick slots
+
 ### Added
 
 - clean up structure to support multiple parallel webviews